+++ src/ts/testingUtils.tsx
@@ -7,6 +7,8 @@ import {render, RenderOptions} from "@testing-library/react";
 // @ts-ignore
 import * as mediaQuery from 'css-mediaquery';
 import {SnackbarProvider} from "notistack";
+import {Grade, Info, Page, Period, Subject, Type} from "../entity";
+import {GradeModalDefaults, NoteRange} from "../entity/config";
 
 const darkTheme = createTheme({
 	palette: {
@@ -69,7 +71,6 @@ const AllTheProviders = ({children}: { children: ReactNode }) => {
 
 const customRender = (ui: ReactElement, options?: Omit<RenderOptions, 'wrapper'>,) => render(ui, {wrapper: AllTheProviders, ...options})
 
-
 function createMatchMedia(width: string): (query: string) => MediaQueryList {
 	return (query: string) => ({
 		matches: mediaQuery.match(query, {width}),
@@ -90,14 +91,37 @@ const rgbToHex = (r: number, g: number, b: number) => '#' + [r, g, b].map(x => {
 	return hex.length === 1 ? '0' + hex : hex
 }).join('')
 
-const convertWeigth = (weight: 'Default' | 'Double' | 'Half') => {
-	switch (weight) {
-		case 'Default':
-			return ""
-		case 'Double':
-			return "x2"
-		case 'Half':
-			return "/2"
+export function mockIPC(args: { periods?: Period[], types?: Type[], subjects?: Subject[], grades?: Grade[], noteRange?: NoteRange, gradeModalDefaults?: GradeModalDefaults, info?: Info, pageFromCache?: Page }) {
+	window.__TAURI_IPC__ = (g) => {
+		switch (g.cmd) {
+			case "get_periods_js":
+				// @ts-ignore
+				return window[`_${g.callback}`](JSON.stringify(args.periods ?? []))
+			case "get_types_js":
+				// @ts-ignore
+				return window[`_${g.callback}`](JSON.stringify(args.types ?? []))
+			case "get_subjects_js":
+				// @ts-ignore
+				return window[`_${g.callback}`](JSON.stringify(args.subjects ?? []))
+			case "get_grades_js":
+				// @ts-ignore
+				return window[`_${g.callback}`](JSON.stringify(args.grades ?? []))
+			case "get_note_range_js":
+				// @ts-ignore
+				return window[`_${g.callback}`](JSON.stringify(args.noteRange ?? {}))
+			case "get_grade_modal_defaults_js":
+				// @ts-ignore
+				return window[`_${g.callback}`](JSON.stringify(args.gradeModalDefaults ?? {}))
+			case "get_info_js":
+				// @ts-ignore
+				return window[`_${g.callback}`](JSON.stringify(args.info ?? {}))
+			case "get_page_from_cache_js":
+				// @ts-ignore
+				return window[`_${g.callback}`](JSON.stringify(args.pageFromCache ?? {}))
+			default:
+				// @ts-ignore
+				window[`_${g.error}`]("Unknown command: " + g.cmd)
+		}
 	}
 }
 
@@ -105,7 +129,6 @@ export {
 	AllTheProviders,
 	rgbToHex,
 	rgbStringToHex,
-	convertWeigth,
 	createMatchMedia
 }
